correcion - mejoras

+++ server/routes/portafolio/servicios.js
@@ -1,7 +1,9 @@
 import express from "express";
 import Servicio from "../../models/portafolio/servicios.js";
 import Categoria from "../../models/categorias.js";
+import Promocion from "../../models/promociones.js";
 import moment from "moment"; // Importa moment para trabajar con fechas
+import { nameDelivery } from "../../utils/varsGlobal.js";
 
 const router = express.Router();
 
@@ -23,7 +25,10 @@ router.post("/add-servicio", (req, res) => {
   newProducto
     .save()
     .then((servicioGuardado) => {
-      res.json(servicioGuardado);
+      res.json({
+        tipoAction: "added",
+        data: servicioGuardado,
+      });
     })
     .catch((error) => {
       console.error("Error al Crear servicio:", error);
@@ -49,7 +54,7 @@ router.get("/get-servicios", async (req, res) => {
         categoria &&
         categoria.name === "Unico" &&
         categoria.nivel === "primario" &&
-        servicio.nombre === "Delivery"
+        servicio.nombre === nameDelivery
       ) {
         // Agregar información del servicio y la categoría al array resultante
         servicioDelivery = servicio;
@@ -78,7 +83,10 @@ router.put("/update-servicio/:idServicio", async (req, res) => {
     );
 
     if (updatedServicio) {
-      return res.json(updatedServicio);
+      return res.json({
+        tipoAction: "updated",
+        data: updatedServicio,
+      });
     } else {
       return res.status(404).json({ mensaje: "No se encontró el servicio" });
     }
@@ -92,9 +100,33 @@ router.delete("/delete-servicio/:idServicio", async (req, res) => {
   const { idServicio } = req.params;
 
   try {
-    const productoEliminado = await Servicio.findByIdAndRemove(idServicio);
-    if (productoEliminado) {
-      return res.json({ mensaje: "Servicio eliminado con éxito" });
+    // Verificar si el servicio está siendo usado en Promociones con alcance distinto de "Todos"
+    const promocionesConServicio = await Promocion.find(
+      { prenda: idServicio, alcance: { $ne: "Todos" } },
+      { _id: 1, codigo: 1 }
+    );
+
+    if (promocionesConServicio.length > 0) {
+      const codigos = promocionesConServicio.map(
+        (promocion) => promocion.codigo
+      );
+      return res.status(400).json({
+        mensaje:
+          "No se puede eliminar el servicio porque está siendo utilizado en una o mas promociones.",
+        codigos,
+      });
+    }
+
+    // Si el servicio no está siendo usado en promociones, procede a eliminarlo
+    const servicioEliminado = await Servicio.findByIdAndRemove(idServicio);
+
+    if (servicioEliminado) {
+      return res.json({
+        tipoAction: "deleted",
+        data: {
+          _id: servicioEliminado._id,
+        },
+      });
     } else {
       return res.status(404).json({ mensaje: "Servicio no encontrado" });
     }
